% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation.R
\name{cite}
\alias{cite}
\title{Citation text}
\usage{
cite(x, ...)
}
\arguments{
\item{x}{A dataset or item from \code{\link[=get_dataset]{get_dataset()}} / \code{\link[=get_item]{get_item()}}. With no
argument, returns the citation for the database itself.}

\item{...}{Unused.}
}
\value{
A character string.
}
\description{
Please cite both the database and the study whose data you used.
}
\examples{
cite()
}
