% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{list_items}
\alias{list_items}
\title{List the stimuli in the database}
\usage{
list_items(version = "latest")
}
\arguments{
\item{version}{Release version, or \code{"latest"}.}
}
\value{
A tibble with one row per item.
}
\description{
List the stimuli in the database
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
list_items()
\dontshow{\}) # examplesIf}
}
