% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{list_studies}
\alias{list_studies}
\title{List the publications in the database}
\usage{
list_studies(version = "latest")
}
\arguments{
\item{version}{Release version, or \code{"latest"}.}
}
\value{
A tibble with one row per study.
}
\description{
List the publications in the database
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
list_studies()
\dontshow{\}) # examplesIf}
}
