% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{release_info}
\alias{release_info}
\title{Version, date and headline counts for the release in use}
\usage{
release_info(version = "latest")
}
\arguments{
\item{version}{Release version, or \code{"latest"}.}
}
\value{
A list describing the release.
}
\description{
Version, date and headline counts for the release in use
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
release_info()
\dontshow{\}) # examplesIf}
}
