% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation.R
\name{bibtex}
\alias{bibtex}
\title{BibTeX for a dataset's source publication}
\usage{
bibtex(x, ...)
}
\arguments{
\item{x}{A dataset from \code{\link[=get_dataset]{get_dataset()}}.}

\item{...}{Unused.}
}
\value{
A character string holding a BibTeX entry.
}
\description{
BibTeX for a dataset's source publication
}
\examples{
\donttest{
bibtex(get_dataset("leeholyoak2021"))
}
}
