% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{clear_cache}
\alias{clear_cache}
\title{Delete cached release files}
\usage{
clear_cache(version = NULL)
}
\arguments{
\item{version}{Version to remove. \code{NULL} (default) clears every version.}
}
\value{
Invisibly, the number of bytes freed.
}
\description{
Delete cached release files
}
\examples{
\donttest{
clear_cache()
}
}
